#2 Break Travel into smaller components
