+++ src/components/Travel/index.js
@@ -1,0 +1,3 @@
+import Travel from './Travel';
+
+export default Travel;
